Refactor Custom Hooks
