--- src/services/getBooks.js
@@ -1,3 +1,0 @@
-export default function getBooks() {
-  return fetch('/books').then((res) => res.json())
-}
